Bump TektonCD Pipeline to v0.41.0
That version is the first LTS following v0.40.0, which fixes this issue
[1] whereby HTTP repo URLs are detected as SSH URLs.

Bumping to the latest version is not an option at this point, as it
removes `git-init` code. There are plans [2] to migrate that code to a
new repository, so we will most likely want to watch that space and
update our build scripts accordingly.

[1]: https://github.com/tektoncd/pipeline/issues/5429
[2]: https://github.com/tektoncd-catalog/git-clone/issues/2